+++ src/components/getFromApi/GetAllMetaData.jsx
@@ -1,4 +1,4 @@
-import axios from './services/AxiosConfig'
+import axios from '../../services/AxiosConfig'
 
 const GetAllMetaData = () => {
     return axios.get('https://hotels4.p.rapidapi.com/get-meta-data')
